+++ src/app/api/counter/route.ts
@@ -1,20 +1,19 @@
 import { NextResponse } from 'next/server';
-import type { NextRequest } from 'next/server';
 
-import { getCurrentCount } from '../quota';
+import { getCurrentCount } from '../counter';
 
 export const dynamic = 'force-dynamic'; // defaults to force-static
 
-export async function GET(request: NextRequest) {
+export async function GET() {
 
   const result = {
     currentCount: await getCurrentCount(),
-  }
+  };
 
   return NextResponse.json(result, {
     headers: {
       // Don't cache
-      //'Cache-Control': 'no-store',
+      'Cache-Control': 'no-cache, no-store',
     }
   });