+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Namaene: Changelog
 
 
+## 0.0.3 (WIP)
+
+- [x] Counter: Refactor to use a hash map
+
+- [x] Counter: Chart (Vega-Lite)
+
+
 ## 0.0.2 (2024-11-02)
 
 - Add favicon.